small refactor for Vec2 and Rect
