+++ rkviewer/canvas/elements.py
@@ -1,0 +1,7 @@
+from abc import abstractmethod
+from .geometry import Vec2
+
+class CanvasElement:
+    @abstractmethod
+    def pos_inside(self, logical_pos: Vec2, ) -> bool:
+        pass
